Corrigindo code smell 'Duplicated Code'
